Fix toroidal edge-wrap to be platform-independent (issue #4)
The toroidal wrap in getCurrentCell only worked by accident of the AVR's
16-bit int. Parameters were `unsigned short` and the wrap was detected with
`if (x == -1)`; on a 16-bit-int AVR, integer promotion makes that compare as
65535 == 65535 (true), but on any 32-bit-int host it becomes 65535 == -1
(false), so the wrap never fires and the framebuffer is read out of bounds.

Make the wrap explicit and width-independent: getCurrentCell/getNeighborCount
now take signed `int` coordinates and wrap with `(i + N) % N` using the
existing ROWS/COLUMNS constants. Signed params are load-bearing -- they ensure
`x - 1` is computed in signed arithmetic and yields a true -1 on both AVR and
host before wrapping. Behavior is identical to the old branches across the
real call domain (offsets -1..25) on AVR, and now also correct on a host.

Add a minimal, test-only host harness so the wrap logic can be exercised off
hardware without the Peggy2Serial library:
- test/Peggy2Serial.h: in-memory grid stub (GetPoint/WritePoint over a bit
  array), confined to test code so it does not pre-empt the larger harness.
- test/test_edge_wrap.cpp: asserts toroidal neighbor counts at corners/edges
  (a single live cell at a corner contributes to the opposite edges).
- test/run.sh: builds with `clang++ -fsanitize=address` and runs.

Verified the original logic FAILS on the host (the three `-1` wraps return 0
instead of 1) and PASSES after the fix (all 10 assertions, clean compile).

+++ Peggy2ConwayEngine.h
@@ -32,8 +32,8 @@ class Peggy2ConwayEngine
 		unsigned short genMemorySize;
 
 		unsigned short isAlive(unsigned short currentState, unsigned short neighborCount);
-		unsigned short getNeighborCount(unsigned short x, unsigned short y);
-		unsigned short getCurrentCell(unsigned short x, unsigned short y);
+		unsigned short getNeighborCount(int x, int y);
+		unsigned short getCurrentCell(int x, int y);
 };
 
 #endif
